requirements.txt
python-gitlab==2.15.0
slack-sdk==3.10.0
